+++ src/sync/converters.test.ts
@@ -1,28 +1,36 @@
-import { parseTimePropertyValue } from '@jupiterone/integration-sdk-core';
 import {
   toOrganizationMemberEntity,
   toOrganizationCollaboratorEntity,
   toRepositoryEntity,
   toAccountEntity,
+  toTeamEntity,
+  toOrgSecretEntity,
+  toRepoSecretEntity,
   toPullRequestEntity,
+  createUnknownUserPrRelationship,
 } from './converters';
-import { UserEntity, PullsListResponseItem } from '../types';
-import { omit } from 'lodash';
+import { UserEntity } from '../types';
 import { PullRequest } from '../client/GraphQLClient/types';
 import {
   fixturePullRequest,
   fixtureUser,
   fixtureReviewerUser,
-  pullRequestReviewer,
-  pullRequestUser,
 } from './fixtures/pullRequest';
-import { toPullRequestEntityOld } from './utils/toPullRequestEntityOld';
 
 describe('toAccountEntity', () => {
   const apiResponse = {
     id: 'account-node-id',
     login: 'account-login',
     name: 'account-name',
+    createdAt: '2021-05-27T15:21:12Z',
+    updatedAt: '2021-09-16T15:04:20Z',
+    description: "Here's my test description",
+    email: 'email@email.com',
+    databaseId: 12345678,
+    isVerified: false,
+    location: 'Transylvania',
+    websiteUrl: 'www.goclickatesturl.com',
+    url: 'https://github.com/SomeOrg',
   };
 
   test('properties transferred', () => {
@@ -42,6 +50,16 @@ describe('toAccountEntity', () => {
       login: 'account-login',
       name: 'account-name',
       displayName: 'account-name',
+      createdOn: 1622128872000,
+      updatedOn: 1631804660000,
+      description: "Here's my test description",
+      email: 'email@email.com',
+      node: 'account-node-id',
+      databaseId: 12345678,
+      isVerified: false,
+      location: 'Transylvania',
+      websiteUrl: 'www.goclickatesturl.com',
+      webLink: 'https://github.com/SomeOrg',
     });
   });
 
@@ -54,6 +72,47 @@ describe('toAccountEntity', () => {
   });
 });
 
+describe('toTeamEntity', () => {
+  const apiResponse = {
+    id: 'MDQ6VGVhbTQ4NTc0OTU=',
+    name: 'Test team',
+    url: 'https://github.com/orgs/Kei-Institute/teams/test-team',
+    slug: 'test-team',
+    createdAt: '2021-06-01T15:45:57Z',
+    updatedAt: '2021-06-01T15:45:57Z',
+    databaseId: 4857495,
+    description: 'Just a test team testing test teams',
+    privacy: 'VISIBLE',
+    organization: 'MDEyOk9yZ2FuaXphdGlvbjg0OTIzNTAz',
+  };
+
+  test('properties transferred', () => {
+    const entity = toTeamEntity(apiResponse as any);
+    //update this
+    expect(entity).toEqual({
+      _class: ['UserGroup'],
+      _type: 'github_team',
+      _key: 'MDQ6VGVhbTQ4NTc0OTU=',
+      _rawData: [
+        {
+          name: 'default',
+          rawData: apiResponse,
+        },
+      ],
+      webLink: 'https://github.com/orgs/Kei-Institute/teams/test-team',
+      name: 'test-team',
+      displayName: 'Test team',
+      fullName: 'Test team',
+      createdOn: 1622562357000,
+      updatedOn: 1622562357000,
+      databaseId: 4857495,
+      description: 'Just a test team testing test teams',
+      node: 'MDQ6VGVhbTQ4NTc0OTU=',
+      privacy: 'VISIBLE',
+    });
+  });
+});
+
 test('toRepositoryEntities', () => {
   const apiResponse = {
     id: 'repo-node-id',
@@ -62,6 +121,26 @@ test('toRepositoryEntities', () => {
     isPrivate: true,
     isArchived: false,
     url: 'web-link',
+    createdAt: '2021-05-27T15:23:24Z',
+    updatedAt: '2021-05-27T15:23:28Z',
+    autoMergeAllowed: false,
+    databaseId: 371419598,
+    deleteBranchOnMerge: false,
+    description: 'This is a test repository',
+    homepageUrl: null,
+    isDisabled: false,
+    isEmpty: false,
+    isFork: false,
+    isInOrganization: true,
+    isLocked: false,
+    isMirror: false,
+    isSecurityPolicyEnabled: false,
+    isTemplate: false,
+    isUserConfigurationRepository: false,
+    lockReason: null,
+    mergeCommitAllowed: true,
+    pushedAt: '2021-06-08T18:12:39Z',
+    rebaseMergeAllowed: true,
   };
   const entity = toRepositoryEntity(apiResponse as any);
   expect(entity).toEqual({
@@ -81,6 +160,27 @@ test('toRepositoryEntities', () => {
     fullName: 'owner/repo-name',
     owner: 'owner',
     archived: false,
+    createdOn: 1622129004000,
+    updatedOn: 1622129008000,
+    pushedOn: 1623175959000,
+    databaseId: 371419598,
+    autoMergeAllowed: false,
+    deleteBranchOnMerge: false,
+    description: 'This is a test repository',
+    homepageUrl: '',
+    node: 'repo-node-id',
+    isDisabled: false,
+    isEmpty: false,
+    isFork: false,
+    isInOrganization: true,
+    isLocked: false,
+    isMirror: false,
+    isSecurityPolicyEnabled: false,
+    isTemplate: false,
+    isUserConfigurationRepository: false,
+    lockReason: '',
+    mergeCommitAllowed: true,
+    rebaseMergeAllowed: true,
   });
 });
 
@@ -92,6 +192,15 @@ describe('toOrganizationMemberEntity', () => {
     hasTwoFactorEnabled: true,
     role: 'Maintainer',
     isSiteAdmin: false,
+    company: 'Test Company',
+    createdAt: '2020-03-22T02:59:22Z',
+    databaseId: 62492097,
+    email: 'flynn@test.com',
+    isEmployee: false,
+    location: 'Best place ever',
+    updatedAt: '2021-08-12T18:55:22Z',
+    url: 'https://github.com/user-login',
+    websiteUrl: 'https://about.me/user-login',
   };
 
   test('properties transferred', () => {
@@ -114,6 +223,15 @@ describe('toOrganizationMemberEntity', () => {
       role: 'Maintainer',
       siteAdmin: false,
       webLink: 'https://github.com/user-login',
+      company: 'Test Company',
+      createdOn: 1584845962000,
+      updatedOn: 1628794522000,
+      databaseId: 62492097,
+      email: 'flynn@test.com',
+      node: 'member-node-id',
+      isEmployee: false,
+      location: 'Best place ever',
+      websiteUrl: 'https://about.me/user-login',
     });
   });
 
@@ -153,6 +271,7 @@ describe('toOrganizationCollaboratorEntity', () => {
       role: 'OUTSIDE',
       siteAdmin: false,
       webLink: 'https://github.com/user-login',
+      node: 'member-node-id',
     });
   });
 
@@ -165,293 +284,71 @@ describe('toOrganizationCollaboratorEntity', () => {
   });
 });
 
-describe('toPullRequestEntity compared against toPullRequestEntityOld', () => {
-  const user = ({
-    name: 'Some Body',
-    displayName: 'Some Body',
-    login: 'somebody',
-    username: 'somebody',
-  } as unknown) as UserEntity;
-  const reviewerUser = ({
-    name: 'Reviewer User',
-    displayName: 'Reviewer User',
-    login: 'reviewer-user',
-    username: 'reviewer-user',
-  } as unknown) as UserEntity;
-  const oldApiResponse: PullsListResponseItem = ({
-    title: 'The Best PR Ever',
-    number: '420',
-    user: {
-      login: 'somebody',
-    },
-    requested_reviewers: [
-      {
-        login: 'reviewer-user',
-      },
-    ],
-    base: {
-      ref: 'qwerty098765',
-      repo: {
-        full_name: 'my-team/my-repo',
-        name: 'my-repo',
-        owner: {
-          login: 'my-team',
+describe('toOrgSecretEntity', () => {
+  const apiResponse = {
+    name: 'KINDA_SECRET',
+    created_at: '2021-08-23T23:42:55Z',
+    updated_at: '2021-08-23T23:42:55Z',
+    visibility: 'selected',
+    selected_repositories_url:
+      'https://api.github.com/orgs/SomeOrg/actions/secrets/KINDA_SECRET/repositories',
+  };
+
+  test('properties transferred', () => {
+    const entity = toOrgSecretEntity(apiResponse as any, 'SomeOrg');
+    expect(entity).toEqual({
+      _class: ['Secret'],
+      _type: 'github_org_secret',
+      _key: 'GitHub_Org_SomeOrg_Secret_KINDA_SECRET',
+      _rawData: [
+        {
+          name: 'default',
+          rawData: apiResponse,
         },
-      },
-    },
-    body: 'This is the description',
-    state: 'MERGED', // Merged because it hase a merged_at property
-    head: {
-      ref: 'abcdef123456',
-    },
-    created_at: '2001-09-11T08:46:00Z',
-    updated_at: '2001-09-11T20:30:00Z',
-    merged_at: '2021-07-21T14:06:13Z',
-    merge_commit_sha: 'f8d8a228a6046ead812d9ea0c457e429342a89f7',
-    url: 'https://api.github.com/repos/JupiterOne/jupiter-project-repo/pulls/1',
-    html_url: 'https://github.com/JupiterOne/jupiter-project-repo/pulls/1',
-  } as unknown) as PullsListResponseItem;
-
-  const newApiResponse: PullRequest = {
-    id: 'id',
-    additions: 1,
-    author: pullRequestUser,
-    authorAssociation: 'authorAssociation',
-    baseRefName: 'qwerty098765',
-    baseRefOid: 'baseRefOid',
-    baseRepository: {
-      name: 'my-repo',
-      owner: {
-        login: 'my-team',
-        id: '12345',
-        url: 'https://github.com/my-team',
-      },
-    },
-    body: 'This is the description',
-    changedFiles: 1,
-    checksUrl: 'checksUrl',
-    closed: false,
-    closedAt: undefined,
-    createdAt: '2001-09-11T08:46:00Z',
-    deletions: 1,
-    editor: pullRequestUser,
-    headRefName: 'abcdef123456',
-    headRefOid: 'headRefOid',
-    headRepository: {
-      name: 'their-repo',
-      owner: {
-        login: 'their-team',
-        id: '67890',
-        url: 'https://github.com/their-team',
-      },
-    },
-    isDraft: false,
-    lastEditedAt: undefined,
-    locked: false,
-    mergeCommit: {
-      oid: 'f8d8a228a6046ead812d9ea0c457e429342a89f7',
-      id: 'id',
-      message: 'message',
-      authoredDate: '2001-09-11T20:30:00Z',
-      changedFiles: 1,
-      commitUrl: 'commitUrl',
-      author: { user: pullRequestUser },
-    },
-    mergeable: 'MERGEABLE',
-    merged: true,
-    mergedAt: '2021-07-21T14:06:13Z',
-    mergedBy: pullRequestUser,
-    number: 420,
-    permalink: 'permalink',
-    publishedAt: undefined,
-    reviewDecision: 'REVIEW_REQUIRED',
-    state: 'MERGED',
-    title: 'The Best PR Ever',
-    updatedAt: '2001-09-11T20:30:00Z',
-    url: 'https://github.com/JupiterOne/jupiter-project-repo/pulls/1',
-    // Optional extra traversals
-    commits: undefined,
-    labels: undefined,
-    reviews: undefined,
-  };
-
-  const expectedEntity = {
-    _type: 'github_pullrequest',
-    _class: ['PR'],
-    _key: `my-team/my-repo/pull-requests/420`,
-    name: 'The Best PR Ever',
-    displayName: `my-repo/420`,
-    id: '420',
-    number: 420,
-    pullRequestId: 'id',
-    accountLogin: 'my-team',
-    authorLogin: 'somebody',
-    author: 'Some Body',
-    reviewerLogins: ['reviewer-user'],
-    reviewers: ['Reviewer User'],
-    approverLogins: ['reviewer-user'],
-    approvers: ['Reviewer User'],
-    title: 'The Best PR Ever',
-    description: `${newApiResponse.body?.substring(0, 80)}...`,
-    state: 'MERGED',
-    source: 'abcdef123456',
-    target: 'qwerty098765',
-    repository: 'my-repo',
-    createdOn: parseTimePropertyValue(newApiResponse.createdAt),
-    updatedOn: parseTimePropertyValue(newApiResponse.updatedAt),
-    webLink: 'https://github.com/JupiterOne/jupiter-project-repo/pulls/1',
-    commits: ['commit_a', 'commit_b', 'commit_c'],
-    commitMessages: ['Commit A', 'Commit B', 'Commit C'],
-    commitsApproved: ['commit_a', 'commit_b'],
-    commitsNotApproved: ['commit_c'],
-    commitsByUnknownAuthor: [],
-    approved: false,
-    allCommitsApproved: false,
-    validated: true,
-    open: false,
-    merged: true,
-    declined: false,
-    mergedOn: 1626876373000,
-    mergeCommitHash: 'f8d8a228a6046ead812d9ea0c457e429342a89f7',
-  };
-
-  test('direct comparison', () => {
-    const oldEntity = toPullRequestEntityOld(
-      oldApiResponse,
-      [
-        { sha: 'commit_a', commit: { message: 'Commit A' } } as any,
-        { sha: 'commit_b', commit: { message: 'Commit B' } } as any,
-        { sha: 'commit_c', commit: { message: 'Commit C' } } as any,
       ],
-      [{ sha: 'commit_a' } as any, { sha: 'commit_b' } as any],
-      [],
-      [{ approverUsernames: ['reviewer-user'] } as any],
-      {
-        [user.login]: user,
-        [reviewerUser.login]: reviewerUser,
-      },
-    );
-    const newEntity = toPullRequestEntity(
-      {
-        ...newApiResponse,
-        commits: [
-          { oid: 'commit_a', message: 'Commit A', author: { user } } as any,
-          { oid: 'commit_b', message: 'Commit B', author: { user } } as any,
-          { oid: 'commit_c', message: 'Commit C', author: { user } } as any,
-        ],
-        reviews: [
-          {
-            id: 'id',
-            state: 'APPROVED',
-            updatedAt: 'updatedAt',
-            url: 'url',
-            author: pullRequestReviewer,
-            commit: {
-              oid: 'commit_b',
-            },
-          },
-        ],
-      },
-      {
-        [user.login]: user,
-        [reviewerUser.login]: reviewerUser,
-      },
-      {
-        [user.login]: user,
-        [reviewerUser.login]: reviewerUser,
-      },
-    );
-
-    // Compare against the previous converter
-    const newPullRequestProperties = [
-      'allCommitsApproved',
-      'pullRequestId',
-      'number',
-    ];
-    expect(omit(newEntity, [...newPullRequestProperties, '_rawData'])).toEqual(
-      omit(oldEntity, ['_rawData']),
-    );
-
-    // Compare with expected
-    expect(omit(newEntity, ['_rawData'])).toEqual(expectedEntity);
-  });
-
-  test('declined comparison', () => {
-    const apiResponseDeclined: PullRequest = {
-      ...newApiResponse,
-      state: 'CLOSED',
-      merged: false,
-      mergedAt: undefined,
-      mergeCommit: undefined,
-    };
-    const newEntity = toPullRequestEntity(
-      {
-        ...apiResponseDeclined,
-        commits: [
-          { oid: 'commit_a', message: 'Commit A', author: { user } } as any,
-          { oid: 'commit_b', message: 'Commit B', author: { user } } as any,
-          { oid: 'commit_c', message: 'Commit C', author: { user } } as any,
-        ],
-        reviews: [
-          {
-            id: 'id',
-            state: 'APPROVED',
-            updatedAt: 'updatedAt',
-            url: 'url',
-            author: pullRequestReviewer,
-            commit: {
-              oid: 'commit_b',
-            },
-          },
-        ],
-      },
-      {
-        [user.login]: user,
-        [reviewerUser.login]: reviewerUser,
-      },
-      {
-        [user.login]: user,
-        [reviewerUser.login]: reviewerUser,
-      },
-    );
-
-    expect(omit(newEntity, ['_rawData'])).toEqual({
-      ...expectedEntity,
-      state: 'CLOSED',
-      open: false,
-      declined: true,
-      merged: false,
-      mergedOn: undefined,
-      mergeCommitHash: undefined,
+      name: 'KINDA_SECRET',
+      displayName: 'KINDA_SECRET',
+      webLink:
+        'https://github.com/organizations/SomeOrg/settings/secrets/actions/KINDA_SECRET',
+      createdOn: 1629762175000,
+      updatedOn: 1629762175000,
+      visibility: 'selected',
+      selectedRepositoriesLink:
+        'https://api.github.com/orgs/SomeOrg/actions/secrets/KINDA_SECRET/repositories',
     });
   });
+});
 
-  test('without commits or reviews', () => {
-    //TODO: fix this
-    const entity = toPullRequestEntity(
-      newApiResponse,
-      {
-        [user.login]: user,
-        [reviewerUser.login]: reviewerUser,
-      },
-      {
-        [user.login]: user,
-        [reviewerUser.login]: reviewerUser,
-      },
+describe('toRepoSecretEntity', () => {
+  const apiResponse = {
+    name: 'PRETTY_SECRET',
+    created_at: '2021-08-23T23:53:00Z',
+    updated_at: '2021-08-23T23:53:00Z',
+  };
+
+  test('properties transferred', () => {
+    const entity = toRepoSecretEntity(
+      apiResponse as any,
+      'SomeOrg',
+      'Test-repo',
     );
-    expect(omit(entity, ['_rawData'])).toEqual({
-      ...expectedEntity,
-      allCommitsApproved: undefined,
-      validated: undefined,
-      commits: undefined,
-      commitMessages: undefined,
-      commitsApproved: undefined,
-      commitsNotApproved: undefined,
-      commitsByUnknownAuthor: undefined,
-      approvers: undefined,
-      approverLogins: undefined,
-      reviewers: undefined,
-      reviewerLogins: undefined,
+    expect(entity).toEqual({
+      _class: ['Secret'],
+      _type: 'github_repo_secret',
+      _key: 'GitHub_Repo_Test-repo_Secret_PRETTY_SECRET',
+      _rawData: [
+        {
+          name: 'default',
+          rawData: apiResponse,
+        },
+      ],
+      name: 'PRETTY_SECRET',
+      displayName: 'PRETTY_SECRET',
+      webLink:
+        'https://github.com/SomeOrg/Test-repo/settings/secrets/actions/PRETTY_SECRET',
+      createdOn: 1629762780000,
+      updatedOn: 1629762780000,
+      visibility: 'selected',
     });
   });
 });
@@ -551,6 +448,34 @@ describe('toPullRequestEntity', () => {
           rawData: fixturePullRequest,
         },
       ],
+    });
+  });
+});
+
+describe('createUnknownUserPrRelationship', () => {
+  test('properties transferred', () => {
+    const relationship = createUnknownUserPrRelationship(
+      'unknownlogin',
+      'github_user_approved_pullrequest',
+      'APPROVED',
+      'somePrEntityKey',
+    );
+    expect(relationship).toEqual({
+      _key: `unknownlogin|approved|somePrEntityKey`,
+      _type: 'github_user_approved_pullrequest',
+      _class: 'APPROVED',
+      _mapping: {
+        sourceEntityKey: 'somePrEntityKey',
+        relationshipDirection: 'REVERSE',
+        targetFilterKeys: [['_type', 'login']],
+        targetEntity: {
+          _class: 'User',
+          _type: 'github_user',
+          login: 'unknownlogin',
+        },
+        skipTargetCreation: false,
+      },
+      displayName: 'APPROVED',
     });
   });
 });