+++ CHANGELOG.md
@@ -8,6 +8,32 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Extra 9 properties to `github_account` (`createdOn`, `updatedOn`,
+  `description`, `email`, `node`, `databaseId`, `isVerified`, `location`,
+  `websiteUrl`, `webLink`)
+- Extra 6 properites to `github_team` (`createdOn`, `updatedOn`, `databaseId`,
+  `description`, `node`, `privacy`)
+- Extra 8 properties to `github_user` (`company`, `createdOn`, `updatedOn`,
+  `databaseId`, `node`, `isEmployee`, `location`, `websiteUrl`, `email`)
+- Extra 19 properties to `github_repo` (`autoMergeAllowed`, `databaseId`,
+  `deleteBranchOnMerge`, `description`, `homepageUrl`, `node`, `isDisabled`,
+  `isEmpty`, `isFork`, `isInOrganization`, `isLocked`, `isMirror`,
+  `isSecurityPolicyEnabled`, `isTemplate`, `isUserConfigurationRepository`,
+  `lockReason`, `mergeCommitAllowed`, `pushedOn`, `rebaseMergeAllowed`)
+- Extra 5 properties to `github_pullrequest` (`databaseId`, `node`,
+  `commitsCount`, `approvalsCount`, `approvalLastAt`)
+- Pull requests Opened, Reviewed, or Approved by a user who is not part of the
+  current organization or collaborator list now have a mapped relationship to a
+  GitHub user with the login recorded in the PR properties.
+
+### Changed
+
+- `createdOn` and `updatedOn` properties for `github_org_secret` and
+  `github_repo_secret` are now time-since-epoch integers instead of strings,
+  matching other entities.
+
 ## 1.4.5 - 2021-09-16
 
 ### Fixed
@@ -17,7 +43,7 @@ and this project adheres to
 
 ### Changed
 
-- To query 25 pull requests at a time insteaad of 50 to prevent Github errors.
+- To query 25 pull requests at a time instead of 50 to prevent Github errors.
 
 ### Added